editoast: move margins tests to schemas
