+++ editoast/src/schema/v2/trainschedule.rs
@@ -194,7 +194,6 @@ pub enum Distribution {
 #[cfg(test)]
 mod tests {
     use chrono::Duration;
-    use editoast_schemas::train_schedule::MarginValue;
     use editoast_schemas::train_schedule::Margins;
     use editoast_schemas::train_schedule::PathItemLocation;
     use editoast_schemas::train_schedule::ScheduleItem;
@@ -203,51 +202,6 @@ mod tests {
 
     use super::PathItem;
     use crate::schema::v2::trainschedule::TrainScheduleBase;
-
-    /// Test that the `MarginValue` enum can be deserialized from a string
-    #[test]
-    fn deserialize_margin_value() {
-        let none: MarginValue = from_str(r#""none""#).unwrap();
-        assert_eq!(none, MarginValue::None);
-
-        let percentage: MarginValue = from_str(r#""10%""#).unwrap();
-        assert_eq!(percentage, MarginValue::Percentage(10.0));
-
-        let min_per_km: MarginValue = from_str(r#""5.3min/km""#).unwrap();
-        assert_eq!(min_per_km, MarginValue::MinPerKm(5.3));
-    }
-
-    /// Test invalid `MarginValue` deserialization
-    #[test]
-    fn deserialize_invalid_margin_value() {
-        assert!(from_str::<MarginValue>(r#""3.5""#).is_err());
-        assert!(from_str::<MarginValue>(r#""-5%""#).is_err());
-        assert!(from_str::<MarginValue>(r#""-0.4min/km""#).is_err());
-    }
-
-    /// Test that the `MarginValue` enum can be serialized to a string
-    #[test]
-    fn serialize_margin_value() {
-        let none = to_string(&MarginValue::None).unwrap();
-        assert_eq!(none, r#""none""#);
-
-        let percentage = to_string(&MarginValue::Percentage(10.0)).unwrap();
-        assert_eq!(percentage, r#""10%""#);
-
-        let min_per_km = to_string(&MarginValue::MinPerKm(5.3)).unwrap();
-        assert_eq!(min_per_km, r#""5.3min/km""#);
-    }
-
-    /// Test that Margins deserialization checks works
-    #[test]
-    fn deserialize_margins() {
-        let valid_margins = r#"{"boundaries":["a", "b"],"values":["none","10%","20min/km"]}"#;
-        assert!(from_str::<Margins>(valid_margins).is_ok());
-        let invalid_margins = r#"{"boundaries":["a"],"values":["none","10%","20min/km"]}"#;
-        assert!(from_str::<Margins>(invalid_margins).is_err());
-        let invalid_margins = r#"{"boundaries":["a", "b"],"values":["none","10%"]}"#;
-        assert!(from_str::<Margins>(invalid_margins).is_err());
-    }
 
     /// Test deserialize a valid train schedule example
     #[test]
